requirements.txt
astroid==2.3.3
backports.ssl-match-hostname==3.7.0.1
boto3==1.12.16
botocore==1.15.16
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
geojson==2.5.0
idna==2.9
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
peewee==3.13.1
psycopg2==2.8.4
pylint==2.4.4
pyowm==2.10.0
python-dateutil==2.8.1
PyYAML==5.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
slackbot==0.5.6
slacker==0.14.0
sseclient==0.0.24
typed-ast==1.4.1
urllib3==1.25.8
websocket-client==0.57.0
wrapt==1.11.2
